# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: RLibrary Backend API
  version: 1.0.0
  description: REST API for managing books, authors, and loans in the RLibrary system.
security:
  - auth0_jwt: []
servers:
  - url: /api
  - url: http://localhost:8080/api
paths:
...
  /persons:
    get:
      summary: Get all persons
      operationId: personsGet
      tags:
        - Person
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            default: lastName
          description: Field to sort by (id, firstName, lastName, birthDate)
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
        - in: query
          name: lastName
          schema:
            type: string
          description: Filter by last name (contains, case-insensitive)
        - in: query
          name: accountStatus
          schema:
            type: string
          description: Filter by account status
      responses:
        '200':
          description: List of persons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPersonResponse'
    post:
      summary: Create a new person
      operationId: personsPost
      tags:
        - Person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonDto'
      responses:
        '201':
          description: Person created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '400':
          description: Invalid birthdate (must be before today)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Library not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /persons/auth0/{auth0Id}:
    get:
      summary: Get person by Auth0 ID
      operationId: personsAuth0Auth0IdGet
      tags:
        - Person
      parameters:
        - name: auth0Id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Person found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '404':
          description: Person not found
  /persons/{id}:
    get:
      summary: Get person by ID
      operationId: personsIdGet
      tags:
        - Person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Person found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '404':
          description: Person not found
    put:
      summary: Update person by ID
      operationId: personsIdPut
      tags:
        - Person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonDto'
      responses:
        '200':
          description: Person updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '400':
          description: Invalid birthdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Person or Library not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete person by ID
      operationId: personsIdDelete
      tags:
        - Person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Person deleted
        '404':
          description: Person not found
components:
  securitySchemes:
    auth0_jwt:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://dev-m362yc372ygvz1ny.us.auth0.com/authorize?audience=https://rlibrary.com
          tokenUrl: https://dev-m362yc372ygvz1ny.us.auth0.com/oauth/token
          scopes:
            openid: OpenId Connect
            profile: User Profile
            email: User Email
  schemas:
...
    PersonDto:
      type: object
      required:
        - firstName
        - lastName
        - birthDate
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        birthDate:
          type: string
          format: date
        functie:
          type: string
          maxLength: 100
        badgeCode:
          type: string
          maxLength: 100
        auth0Id:
          type: string
          readOnly: true
          description: Auth0 subject ID (e.g. "auth0|abc123"). Set at registration, never changed by the user.
        userRole:
          $ref: '#/components/schemas/UserRole'
        gender:
          $ref: '#/components/schemas/Gender'
        country:
          type: string
          maxLength: 100
          description: Country of residence, e.g. "Belgium"
        phoneNumber:
          type: string
          maxLength: 20
        preferredLanguage:
          type: string
          maxLength: 10
          description: Preferred UI language, e.g. "nl", "en", "fr"
        profilePictureUrl:
          type: string
          maxLength: 500
          description: URL to the user's profile picture
        activeLibraryId:
          type: integer
          format: int64
        accountStatus:
          type: string
        personalStatsId:
          type: integer
          format: int64
        loanIds:
          type: array
          readOnly: true
          items:
            type: integer
            format: int64
        notificationIds:
          type: array
          readOnly: true
          items:
            type: integer
            format: int64
        bookComplaintIds:
          type: array
          readOnly: true
          items:
            type: integer
            format: int64
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
        message:
          type: string
...
    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
        - OTHER
        - PREFER_NOT_TO_SAY
    UserRole:
      type: string
      description: Role of the user in the application. Mirrors Auth0 roles. NOT changeable by the user themselves.
      enum:
        - WERKNEMER
        - VERANTWOORDELIJKE
...
    PagedPersonResponse:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
        - items
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/PersonDto'